refactored python interface files a bit
