+++ py/MANIFEST.in
@@ -1,0 +1,5 @@
+include MANIFEST.in
+include *.txt
+
+# Header files
+include cxx/tomographerpy/*.h
