+++ test/pytests.py
@@ -7,7 +7,7 @@ import os.path
 print("Run: "+" ".join(sys.argv))
 
 TomographerPyMod = sys.argv[1]
-sys.path.append(os.path.dirname(TomographerPyMod))
+sys.path = [ os.path.dirname(TomographerPyMod) ] + sys.path
 
 # import the module
 import tomographer
@@ -25,5 +25,10 @@ print("\n\nNOT YET IMPLEMENTED: TODO: WRITE TESTS FOR THE TOMOGRAPHER PYTHON INT
 
 # it's ugly on travis if the whole test suite fails because of unwritten tests...
 # only report failure when running the tests locally
-if 'TRAVIS_OS' not in os.environ:
+if 'TRAVIS' not in os.environ:
     sys.exit(1)
+
+
+
+
+sys.exit(0)
